Refactor the code to make it less dependable
* Make play optional
* Support Play 2.4.8 and 2.6.6
* Make metrics optional
* Support dropwizard
