--- .java-version
@@ -1,1 +1,0 @@
-oracle64-1.8.0.11
