--- circle.yml
@@ -1,10 +1,0 @@
-machine:
-  java:
-    version: oraclejdk8
-
-test:
-  override:
-    - (cd module-code; sbt clean coverage test)
-    - (cd module-code; sbt coverageReport)
-    - (cd module-code; sbt coverageAggregate)
-    - (cd module-code; sbt codacyCoverage)
