--- module-code/app/com/codacy/play/tomcatcp/metrics/MetricsTrackerFactory.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) 2013,2014 Brett Wooldridge
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.codacy.play.tomcatcp.metrics;
-
-public interface MetricsTrackerFactory {
-    /**
-     * Create an instance of a MetricsTracker.
-     *
-     * @param poolName  the name of the pool
-     * @param poolStats a PoolStats instance to use
-     * @return a MetricsTracker implementation instance
-     */
-    MetricsTracker create(String poolName, PoolStats poolStats);
-}
