--- module-code/project/plugins.sbt
@@ -1,25 +1,0 @@
-import sbt._
-
-ivyLoggingLevel := UpdateLogging.Full
-
-// Comment to get more information during initialization
-logLevel := Level.Info
-
-resolvers ++= Seq(
-    DefaultMavenRepository,
-    "Typesafe Repository" at "http://repo.typesafe.com/typesafe/releases/",
-    "Sonatype snapshots" at "http://oss.sonatype.org/content/repositories/snapshots/",
-    Classpaths.typesafeReleases
-)
-
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % System.getProperty("play.version", "2.6.2"))
-
-addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "1.1")
-
-addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
-
-addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.1.8")
-
-addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.3.3")
-
-addSbtPlugin("com.codacy" % "sbt-codacy-coverage" % "1.2.1")
