+++ project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=0.13.16
