+++ project/plugins.sbt
@@ -1,0 +1,23 @@
+import sbt._
+
+ivyLoggingLevel := UpdateLogging.Full
+
+// Comment to get more information during initialization
+logLevel := Level.Info
+
+resolvers ++= Seq(
+  DefaultMavenRepository,
+  "Typesafe Repository" at "http://repo.typesafe.com/typesafe/releases/",
+  "Sonatype snapshots" at "http://oss.sonatype.org/content/repositories/snapshots/",
+  Classpaths.typesafeReleases
+)
+
+addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "2.0")
+
+addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.1.0")
+
+addSbtPlugin("com.timushev.sbt" % "sbt-updates" % "0.3.2")
+
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.1")
+
+addSbtPlugin("com.codacy" % "sbt-codacy-coverage" % "1.3.11")
